% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_selected_filters.R
\name{selected_filters_ui}
\alias{selected_filters_ui}
\alias{selected_filters_server}
\title{selected_filters_ui and selected_filters_server}
\usage{
selected_filters_ui(id)

selected_filters_server(
  input,
  output,
  session,
  group_inputs,
  label_none = "None",
  labels = c()
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{group_inputs}{From shinyWidgets \code{pickerGroup} or \code{selectizeGroup}}

\item{label_none}{Label display if no filter is set}

\item{labels}{Rename filter codes to labels in sidebar}
}
\description{
A shiny Module.
}
